product browse and admin endpoints configured

+++ src/main/java/com/app/ecom_app/controller/AdminUserController.java
@@ -1,21 +1,21 @@
 package com.app.ecom_app.controller;
 
-import com.app.ecom_app.dto.UserRequest;
 import com.app.ecom_app.dto.UserResponse;
 import com.app.ecom_app.enums.UserRole;
 import com.app.ecom_app.model.User;
 import com.app.ecom_app.repository.UserRepo;
 import com.app.ecom_app.service.UserService;
-import com.app.ecom_app.utils.UserMappers;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("/admin")
+@RequestMapping("/admin/users")
 @RequiredArgsConstructor
 public class AdminUserController {
 
@@ -25,6 +25,7 @@ public class AdminUserController {
 
     @GetMapping
     public ResponseEntity<List<UserResponse>> getAllUsers() {
+
         List<UserResponse> users = userService.fetchAllUsers();
         if (users.isEmpty()) {
             return new ResponseEntity<>(users, HttpStatus.NO_CONTENT);
@@ -34,6 +35,7 @@ public class AdminUserController {
 
     @GetMapping("/{id}")
     public ResponseEntity<UserResponse> getUserById(@PathVariable String id) {
+
         UserResponse user = userService.fetchUserById(id);
         if (user == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -43,6 +45,7 @@ public class AdminUserController {
 
     @PutMapping("/{id}")
     public ResponseEntity<UserResponse> updateToAdminRole(@PathVariable String id) {
+
         User user = userRepo.findById(id).orElse(null);
         if (user == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -52,9 +55,13 @@ public class AdminUserController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<UserResponse> deleteUserById(@PathVariable String id) {
+
         User user = userRepo.findById(id).orElse(null);
         if (user == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        if(user.getRole() == UserRole.ADMIN){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
         userRepo.delete(user);
         return new ResponseEntity<>(HttpStatus.OK);
